+++ src/app/folders/@common/components/AlertNoData/AlertNoData.module.scss
@@ -1,0 +1,12 @@
+.component{
+  :global(.MuiAlert-icon) {
+    font-size: 18px;
+  }
+}
+
+.link{
+  cursor: pointer;
+  text-decoration: underline;
+  font-size: 1.02rem;
+  font-weight: 500;
+}
