+++ src/app/folders/index.tsx
@@ -3,12 +3,12 @@ import Grid from '@material-ui/core/Grid';
 
 import SimpleConfirmation from 'lib/modals/SimpleConfirmation';
 import GeneralLayout from 'app/system/layout';
-import Text from 'lib/text';
 
 import useRemoveFolder from './@common/hooks/removeFolder';
 import useFetchFolders from './@common/hooks/fetchFolders';
 import ActionBarFolders from './@common/components/ActionBar';
 import TableView from './TableView';
+import AlertNoData from './@common/components/AlertNoData';
 
 const ListFoldersPage: React.FC = () => {
   const { fetch, list, loading, noData } = useFetchFolders();
@@ -26,11 +26,7 @@ const ListFoldersPage: React.FC = () => {
       </Grid>
       <Grid container justifyContent="center" spacing={2}>
         <Grid item xs={11}>
-          {noData && (
-            <Text textAlign="center" variant="h6">
-              There are no folders yet
-            </Text>
-          )}
+          {noData && <AlertNoData />}
           {!noData && (
             <TableView
               list={list}
