+++ src/common/@types/general.ts
@@ -1,1 +1,5 @@
+import React from 'react';
+
 export type Filter<T> = (item: T, index: number, arr: T[]) => boolean;
+
+export type EventHandler = (event: React.MouseEvent<HTMLButtonElement>) => void;
